added support for figures to questions. closes #7.
The Question class can now store figures. The LaTeX and Blackboard
writers will look to see if a question has a figure and insert it
accordingly. The LaTeX writer simply adds it to the list of figures that
will be added. The Blackboard writer will embed it in the question text.