fix: fixed bug in latex quiz key generator
The latex quiz key generator was not working for some questions. It
seemed like it didn't work when the answer contained latex.

The problem turned out to be that we were taking the id() of the
formatted answer, which is generated on-demand, so the text string that
was returned had a different id each time.

If an answer doesn't have any formatting tokens, then it will just
be returned directly, so only the answers that contained {} (i.e. latex)
were suffering from this problem.