fix: Blackboard writer strips new lines from question text.
The text files uploaded by Blackboard use newlines to delimit separate
questions, so the question text cannot have a newline character in it.
This is inconvenient because you can't format the question text with
line breaks to make it more readable.

This commit fixes the issue. You can now format question text for
readability, and the Blackboard writer will strip out line breaks an
multiple spaces.